+++ packages/ui/src/cairo_alpha/ERC721Controls.svelte
@@ -2,12 +2,13 @@
   import HelpTooltip from '../common/HelpTooltip.svelte';
 
   import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo-alpha';
-  import { erc721, infoDefaults } from '@openzeppelin/wizard-cairo-alpha';
+  import { erc721, infoDefaults, macrosDefaults } from '@openzeppelin/wizard-cairo-alpha';
 
   import AccessControlSection from './AccessControlSection.svelte';
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import RoyaltyInfoSection from './RoyaltyInfoSection.svelte';
   import InfoSection from './InfoSection.svelte';
+  import MacrosSection from './MacrosSection.svelte';
   import ExpandableCheckbox from '../common/ExpandableCheckbox.svelte';
   import { error } from '../common/error-tooltip';
 
@@ -16,6 +17,8 @@
     ...erc721.defaults,
     royaltyInfo: { ...erc721.defaults.royaltyInfo }, // copy fields
     info: { ...infoDefaults }, // create new object since Info is nested
+    macros: { ...macrosDefaults }, // create new object since MacrosOptions is nested
+    access: { ...erc721.defaults.access }, // create new object since Access is nested
   };
 
   export let errors: undefined | OptionsErrorMessages;
@@ -52,7 +55,7 @@
     <label class:checked={opts.mintable}>
       <input type="checkbox" bind:checked={opts.mintable} />
       Mintable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/erc721">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/alpha/erc721">
         Privileged accounts will be able to emit new tokens.
       </HelpTooltip>
     </label>
@@ -64,7 +67,7 @@
     <label class:checked={opts.pausable}>
       <input type="checkbox" bind:checked={opts.pausable} />
       Pausable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/security#pausable">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/alpha/security#pausable">
         Privileged accounts will be able to pause the functionality marked with <code
           >self.pausable.assert_not_paused()</code
         >. Useful for emergency response.
@@ -73,7 +76,7 @@
     <label class:checked={opts.enumerable}>
       <input type="checkbox" bind:checked={opts.enumerable} />
       Enumerable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/api/erc721#ERC721EnumerableComponent">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/alpha/api/erc721#ERC721EnumerableComponent">
         Allows a contract to publish its entire list of NFTs and make them discoverable by keeping track of all token
         ids and all tokens owned by an address.
       </HelpTooltip>
@@ -86,13 +89,13 @@
   label="Votes"
   bind:checked={opts.votes}
   helpContent="Keeps track of historical balances for voting in on-chain governance, with a way to delegate one's voting power to a trusted account."
-  helpLink="https://docs.openzeppelin.com/contracts-cairo/governance/votes"
+  helpLink="https://docs.openzeppelin.com/contracts-cairo/alpha/governance/votes"
   error={errors?.appName || errors?.appVersion}
 >
   <label class="labeled-input">
     <span class="flex justify-between pr-2">
       Application Name
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/guides/snip12">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/alpha/guides/snip12">
         Name for domain separator. Prevents two applications from producing the same hash.
       </HelpTooltip>
     </span>
@@ -102,7 +105,7 @@
   <label class="labeled-input">
     <span class="flex justify-between pr-2">
       Application Version
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/guides/snip12">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/alpha/guides/snip12">
         Version for domain separator. Prevents two versions of the same application from producing the same hash.
       </HelpTooltip>
     </span>
@@ -112,6 +115,14 @@
 
 <RoyaltyInfoSection bind:opts={opts.royaltyInfo} {errors} />
 
-<AccessControlSection bind:access={opts.access} required={requireAccessControl} />
+<AccessControlSection
+  bind:accessType={opts.access.type}
+  bind:darInitialDelay={opts.access.darInitialDelay}
+  bind:darDefaultDelayIncrease={opts.access.darDefaultDelayIncrease}
+  required={requireAccessControl}
+  {errors}
+/>
+
+<MacrosSection bind:macros={opts.macros} />
 
 <InfoSection bind:info={opts.info} />